ServiceController is now abstract as this class should not be instantiated.
Also, the provideRetrofit method in RestServiceModule was modified to receive a Converter.Factory. Before this change, it received a ServiceStringConverter which restricted you to use the existing implementation or to Override it. You can now use the converter you wish, as long as it is of the Converter.Factory type.